Api internal functions

Primitive Api wrapper functions are not exported since they only allow for one call at a time. Instead, an equivalent functions is exposed allowing multiple parameters, making it easier to retrieve data for users.

Arguments

competition_code

One or more competition codes.
Admitted values are E for Euroleague and U for Eurocup.

team_code

One or more team codes as obtained from getCompetitionTeams().
Examples are ASV, MAD, ...

game_code

One or more game codes as obtained from getCompetitionGames().

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

round

One or more round codes as obtained from getCompetitionRounds().

phase_type

One or more phase type codes as obtained from getCompetitionPhases().
Admitted values are RS for regular season, PO for playoffs and FF for final four. Default is All for all.

subset

One or more game subsets.
Admitted values are HomeGames, AwayGames, GamesWon, GamesLost, ResultsIn5Points (for games resulted in +/-5 points) and All. Default is All.

statistic_mode

One or more aggregation modes of statistics.
Admitted values are perGame, perMinute and accumulated.