Get competition calendar standings

Get competition calendar standings

Usage

getCompetitionCalendar(season_code, round)

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

round

One or more round codes as obtained from getCompetitionRounds().

Value

Returns a summary tibble of calendar standings for chosen competitions and rounds

Reference webpage: Calendar standings

Examples

if(interactive()) {

getCompetitionCalendar(season_code = c("E2023", "E2022"), round = 1)

}