Get competition history

Usage

getCompetitionHistory(competition_code)

Arguments

competition_code

One or more competition codes.
Admitted values are E for Euroleague and U for Eurocup.

Value

Returns a summary tibble of history for chosen competitions

Examples

getCompetitionHistory(competition_code = c("E", "U"))
#> # A tibble: 46 × 15
#>    CompetitionCode Name  SeasonCode Alias  Year StartDate ActivationDate EndDate
#>    <chr>           <chr> <chr>      <chr> <int> <chr>     <chr>          <chr>  
#>  1 E               Euro… E2023      2023…  2023 2023-06-… 2023-06-29T00… 2024-0…
#>  2 E               Euro… E2022      2022…  2022 2022-07-… 2022-07-01T00… 2023-0…
#>  3 E               Euro… E2021      2021…  2021 2021-09-… 2021-07-01T00… 2022-0…
#>  4 E               Euro… E2020      2020…  2020 2020-09-… 2020-06-23T00… 2021-0…
#>  5 E               Euro… E2019      2019…  2019 2019-09-… 2019-07-01T00… 2020-0…
#>  6 E               Euro… E2018      2018…  2018 2018-09-… 2018-07-02T00… 2019-0…
#>  7 E               Euro… E2017      2017…  2017 2017-10-… 2017-07-01T00… 2018-0…
#>  8 E               Euro… E2016      2016…  2016 2016-09-… 2016-07-01T00… 2017-0…
#>  9 E               Euro… E2015      2015…  2015 2015-09-… 2015-08-01T00… 2016-0…
#> 10 E               Euro… E2014      2014…  2014 2014-09-… 2014-08-01T00… 2015-0…
#> # ℹ 36 more rows
#> # ℹ 7 more variables: WinnerCode <chr>, WinnerName <chr>,
#> #   WinnerAbbreviatedName <chr>, WinnerEditorialName <chr>, WinnerTvCode <chr>,
#> #   WinnerIsVirtual <lgl>, WinnerImages <chr>