Competition metadata

Retrieve values of arguments for specific data collection functions across all package.

Usage

getCompetitionHistory(competition_code)

getCompetitionRounds(season_code)

getCompetitionPhases(season_code)

getCompetitionTeams(season_code)

getCompetitionGames(season_code, round, phase_type = "All")

Arguments

competition_code

One or more competition codes.
Admitted values are E for Euroleague and U for Eurocup.

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

round

One or more round codes as obtained from getCompetitionRounds().

phase_type

One or more phase type codes as obtained from getCompetitionPhases().
Admitted values are RS for regular season, PO for playoffs and FF for final four. Default is All for all.

Examples

getCompetitionHistory(competition_code = c("E", "U")) |> head(5)
#> # A tibble: 5 × 15
#>   CompetitionCode Name          SeasonCode Alias  Year StartDate  ActivationDate
#>   <chr>           <chr>         <chr>      <chr> <int> <date>     <date>        
#> 1 E               EuroLeague 2… E2023      2023…  2023 2023-06-29 2023-06-29    
#> 2 E               EuroLeague 2… E2022      2022…  2022 2022-07-01 2022-07-01    
#> 3 E               EuroLeague 2… E2021      2021…  2021 2021-09-01 2021-07-01    
#> 4 E               EuroLeague 2… E2020      2020…  2020 2020-09-01 2020-06-23    
#> 5 E               EuroLeague 2… E2019      2019…  2019 2019-09-01 2019-07-01    
#> # ℹ 8 more variables: EndDate <date>, WinnerCode <chr>, WinnerName <chr>,
#> #   WinnerAbbreviatedName <chr>, WinnerEditorialName <chr>, WinnerTvCode <chr>,
#> #   WinnerIsVirtual <lgl>, WinnerImages <chr>

getCompetitionRounds(season_code = c("E2023", "E2022")) |> head(5)
#> # A tibble: 5 × 8
#>   SeasonCode PhaseTypeCode Round Index Name    MinGameStartDate MaxGameStartDate
#>   <chr>      <chr>         <int> <int> <chr>   <date>           <date>          
#> 1 E2023      RS               34    34 Round … 2024-04-08       2024-04-14      
#> 2 E2023      RS               33    33 Round … 2024-04-01       2024-04-07      
#> 3 E2023      RS               32    32 Round … 2024-03-25       2024-03-31      
#> 4 E2023      RS               31    31 Round … 2024-03-21       2024-03-24      
#> 5 E2023      RS               30    30 Round … 2024-03-18       2024-03-20      
#> # ℹ 1 more variable: DatesFormmated <chr>

getCompetitionPhases(season_code = c("E2023", "U2023")) |> head(5)
#> # A tibble: 3 × 15
#>   SeasonCode SeasonName         SeasonAlias SeasonCompetitionCode SeasonYear
#>   <chr>      <chr>              <chr>       <chr>                      <int>
#> 1 E2023      EuroLeague 2023-24 2023-24     E                           2023
#> 2 U2023      EuroCup 2023-24    2023-24     U                           2023
#> 3 U2023      EuroCup 2023-24    2023-24     U                           2023
#> # ℹ 10 more variables: SeasonStartDate <date>, PhaseType <chr>,
#> #   PhaseTypeAlias <chr>, PhaseTypeName <chr>, PhaseTypeIsGroupPhase <lgl>,
#> #   StartDate <date>, EndDate <date>, Order <int>, HasPlayedGames <lgl>,
#> #   HasGames <lgl>

getCompetitionTeams(season_code = c("E2023", "U2023")) |> head(5)
#> # A tibble: 5 × 25
#>   SeasonCode TeamCode TeamName      AbbreviatedName TvCode IsVirtual ImagesCrest
#>   <chr>      <chr>    <chr>         <chr>           <chr>  <lgl>     <chr>      
#> 1 E2023      BER      ALBA Berlin   ALBA            BER    FALSE     https://me…
#> 2 E2023      IST      Anadolu Efes… Efes            EFS    FALSE     https://me…
#> 3 E2023      MCO      AS Monaco     Monaco          ASM    FALSE     https://me…
#> 4 E2023      BAS      Baskonia Vit… Baskonia        BKN    FALSE     https://me…
#> 5 E2023      RED      Crvena Zvezd… Zvezda          CZV    FALSE     https://me…
#> # ℹ 18 more variables: ImagesOnDarkCrest <chr>, ImagesOnLightCrest <chr>,
#> #   EditorialName <chr>, Sponsor <chr>, ClubPermanentName <chr>,
#> #   ClubPermanentAlias <chr>, CountryCode <chr>, CountryName <chr>,
#> #   Address <chr>, Website <chr>, TicketsUrl <chr>, TwitterAccount <chr>,
#> #   VenueCode <chr>, City <chr>, President <chr>, Phone <chr>,
#> #   PrimaryColor <chr>, SecondaryColor <chr>

getCompetitionGames(season_code = "E2023", round = 1:5) |> head(5)
#> # A tibble: 5 × 78
#>   SeasonCode Round PhaseType Id       GameCode SeasonName SeasonAlias SeasonYear
#>   <chr>      <int> <chr>     <chr>       <int> <chr>      <chr>            <int>
#> 1 E2023          1 All       0f3e055…        5 EuroLeagu… 2023-24           2023
#> 2 E2023          1 All       57eeec3…        2 EuroLeagu… 2023-24           2023
#> 3 E2023          1 All       68ea1f2…        3 EuroLeagu… 2023-24           2023
#> 4 E2023          1 All       7f6a4bb…        9 EuroLeagu… 2023-24           2023
#> 5 E2023          1 All       9a5a317…        8 EuroLeagu… 2023-24           2023
#> # ℹ 70 more variables: CompetitionCode <chr>, CompetitionName <chr>,
#> #   GroupId <chr>, GroupName <chr>, GroupOrder <int>, PhaseTypeCode <chr>,
#> #   PhaseTypeName <chr>, PhaseTypeAlias <chr>, PhaseTypeIsGroupPhase <lgl>,
#> #   RoundName <chr>, RoundAlias <chr>, Date <date>, Status <chr>, Minute <chr>,
#> #   RemainingTime <chr>, Quarter <lgl>, QuarterMinute <lgl>, HomeCode <chr>,
#> #   HomeName <chr>, HomeAbbreviatedName <chr>, HomeTla <chr>, HomeScore <int>,
#> #   HomeStandingsScore <int>, HomeQuartersQ1 <int>, HomeQuartersQ2 <int>, …