Get game box-score

Usage

getGameBoxScore(game_code, season_code)

Arguments

game_code

One or more game codes as obtained from getCompetitionGames().

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

Value

Returns a list of elements for the chosen games and seasons

Team. Name of the teams

Coach. Name of the coaches

EndOfQuarter. Team accumulated points by quarter

ByQuarter. Team points for each quarter

PlayerStats. Statistics for each player in the game

TeamStats. Aggregated statistics for each team in the game

Glossary of columns:

Column name	Column extended name

GP	Game player

GS	Game started

MIN	Minutes played

PTS	Points scored

2PM	Two-pointers made

2PA	Two-pointers attempted

2P%	Two-point %

3PM	Three-pointers made

3PA	Three-pointers attempted

3P%	Three-point %

FTM	Free throws made

FTA	Free throws attempted

FT%	Free-throw %

OREB	Offensive rebounds

DREB	Defensive rebounds

TREB	Total rebounds

AST	Assists

STL	Steals

TO	Turnovers

BLK	Blocks

BLKA	Blocks against

FC	Personal fouls committed

FD	Personal fouls drawn

PIR	Performance Index Rating

Reference webpage: BoxScore

Examples

getGameBoxScore(season_code = c("E2023", "U2023"), game_code = 1)
#> $Team
#> # A tibble: 4 × 3
#>   GameCode SeasonCode Team                              
#>      <dbl> <chr>      <chr>                             
#> 1        1 E2023      CRVENA ZVEZDA MERIDIANBET BELGRADE
#> 2        1 E2023      LDLC ASVEL VILLEURBANNE           
#> 3        1 U2023      PROMETEY SLOBOZHANSKE             
#> 4        1 U2023      VEOLIA TOWERS HAMBURG             
#> 
#> $Coach
#> # A tibble: 4 × 3
#>   GameCode SeasonCode Coach            
#>      <dbl> <chr>      <chr>            
#> 1        1 E2023      IVANOVIC, DUSKO  
#> 2        1 E2023      PARKER, TJ       
#> 3        1 U2023      GINZBURG, RONEN  
#> 4        1 U2023      BARLOSCHKY, BENKA
#> 
#> $EndOfQuarter
#> # A tibble: 4 × 7
#>   GameCode SeasonCode Team                   Quarter1 Quarter2 Quarter3 Quarter4
#>      <dbl> <chr>      <chr>                     <int>    <int>    <int>    <int>
#> 1        1 E2023      CRVENA ZVEZDA MERIDIA…       29       52       69       94
#> 2        1 E2023      LDLC ASVEL VILLEURBAN…       18       31       55       73
#> 3        1 U2023      PROMETEY SLOBOZHANSKE        35       54       78       97
#> 4        1 U2023      VEOLIA TOWERS HAMBURG        15       42       62       77
#> 
#> $ByQuarter
#> # A tibble: 4 × 7
#>   GameCode SeasonCode Team                   Quarter1 Quarter2 Quarter3 Quarter4
#>      <dbl> <chr>      <chr>                     <int>    <int>    <int>    <int>
#> 1        1 E2023      CRVENA ZVEZDA MERIDIA…       29       23       17       25
#> 2        1 E2023      LDLC ASVEL VILLEURBAN…       18       13       24       18
#> 3        1 U2023      PROMETEY SLOBOZHANSKE        35       19       24       19
#> 4        1 U2023      VEOLIA TOWERS HAMBURG        15       27       20       15
#> 
#> $PlayerStats
#> # A tibble: 45 × 31
#>    GameCode SeasonCode Player_ID IsStarter IsPlaying TeamCode Player     Seconds
#>       <dbl> <chr>      <chr>         <int>     <int> <chr>    <chr>        <dbl>
#>  1        1 E2023      004720            0         1 RED      STEFAN LA…     304
#>  2        1 E2023      JDR               0         1 RED      MILOS TEO…     877
#>  3        1 E2023      005099            0         1 RED      ADAM HANG…    1341
#>  4        1 E2023      002553            0         1 RED      LUKA MITR…    1292
#>  5        1 E2023      LUO               1         0 RED      BRANKO LA…     766
#>  6        1 E2023      012711            1         0 RED      MARKO SIM…    1102
#>  7        1 E2023      011224            1         0 RED      SHABAZZ N…    1240
#>  8        1 E2023      008966            0         0 RED      MIKE TOBE…     784
#>  9        1 E2023      011964            0         1 RED      DALIBOR I…     873
#> 10        1 E2023      000848            0         0 RED      NEMANJA N…    1017
#> # ℹ 35 more rows
#> # ℹ 23 more variables: PTS <int>, `2PM` <int>, `2PA` <int>, `3PM` <int>,
#> #   `3PA` <int>, FTM <int>, FTA <int>, OREB <int>, DREB <int>, REB <int>,
#> #   AST <int>, STL <int>, TO <int>, BLK <int>, BLKA <int>, FC <int>, FD <int>,
#> #   PIR <int>, PM <int>, `FG%` <dbl>, `2P%` <dbl>, `3P%` <dbl>, `FT%` <dbl>
#> 
#> $TeamStats
#> # A tibble: 4 × 26
#>   GameCode SeasonCode TeamCode Seconds   PTS `2PM` `2PA` `3PM` `3PA`   FTM   FTA
#>      <dbl> <chr>      <chr>      <dbl> <int> <int> <int> <int> <int> <int> <int>
#> 1        1 E2023      RED        12000    94    16    37    17    34    11    13
#> 2        1 E2023      ASV        12000    73    17    29     9    25    12    21
#> 3        1 U2023      PMT        12000    97    23    33    12    29    15    20
#> 4        1 U2023      HAM        12000    77    16    38    11    36    12    23
#> # ℹ 15 more variables: OREB <int>, DREB <int>, REB <int>, AST <int>, STL <int>,
#> #   TO <int>, BLK <int>, BLKA <int>, FC <int>, FD <int>, PIR <int>,
#> #   `FG%` <dbl>, `2P%` <dbl>, `3P%` <dbl>, `FT%` <dbl>
#>