Get game evolution

Usage

getGameEvolution(game_code, season_code)

Arguments

game_code

The code of the game as obtained from getCompetitionGames().

season_code

The code of the season as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

Value

Returns a list of two elements for the chosen games and seasons

EvolutionSummary. Overall information about minimum and maximum difference of scores between teams

Evolution. Minute by minute points of each team

Reference webpage: GraphicStats

Examples

getGameEvolution(season_code = c("E2023", "U2023"), game_code = 1)
#> $EvolutionSummary
#> # A tibble: 2 × 6
#>   GameCode SeasonCode MinuteMaxA MinuteMaxB ScoreMaxA ScoreMaxB
#>      <dbl> <chr>      <chr>      <chr>      <chr>     <chr>    
#> 1        1 E2023      23         0          62 - 37   0 - 0    
#> 2        1 U2023      11         0          38 - 15   0 - 0    
#> 
#> $Evolution
#> # A tibble: 82 × 7
#>    GameCode SeasonCode Minute PointsTeamA PointsTeamB DiffTeamA DiffTeamB
#>       <dbl> <chr>       <int>       <int>       <int>     <int>     <int>
#>  1        1 E2023           0           3           2         1        NA
#>  2        1 E2023           1           6           4         2        NA
#>  3        1 E2023           2          12           6         6        NA
#>  4        1 E2023           3          17           7        10        NA
#>  5        1 E2023           4          19          10         9        NA
#>  6        1 E2023           5          22          14         8        NA
#>  7        1 E2023           6          24          14        10        NA
#>  8        1 E2023           7          26          16        10        NA
#>  9        1 E2023           8          29          16        13        NA
#> 10        1 E2023           9          29          18        11        NA
#> # ℹ 72 more rows
#>