Get player advanced statistics

Usage

getPlayerAdvanced(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

statistic_mode

One or more aggregation modes of statistics.
Admitted values are perGame, perMinute and accumulated.

Value

Returns a summary tibble of advanced players statistics for chosen seasons.

Glossary of columns:

Column name	Column extended name	Column description

eFG%	Effective field goal %	Combined two- and three-point shooting effectiveness

TS%	True shooting %	Percentage of points vs. points attempted

OREB%	Offensive rebound %	Estimated % of available offensive rebounds obtained while on court

DREB%	Defensive rebound %	Estimated % of available defensive rebounds obtained while on court

REB%	Rebound %	Estimated % of available rebounds obtained while on court

AST/TO	Assist to turnover ratio	Ratio of assists made to turnovers committed

AST-R	Assist ratio	Estimated % of assists per player's offensive possessions

TO-R	Turnover ratio	Estimated % of turnovers per player's offensive possessions

2PTA-R	Two-point attempts ratio	Estimated % of two-point attempts per player's offensive possessions

3PTA-R	Three-point attempts ratio	Estimated % of three-point attempts per player's offensive possessions

FT-RT	Free Throw rate	Measure of free throw attempts vs. field goal attempts

Reference webpage: Stats

Examples

getPlayerAdvanced(season_code = "E2023", statistic_mode = "perGame")
#> # A tibble: 210 × 24
#>    SeasonCode StatisticMode Player_ID PlayerName     PlayerAge ImageUrl TeamCode
#>    <chr>      <chr>         <chr>     <chr>              <int> <chr>    <chr>   
#>  1 E2023      perGame       007200    LARKIN, SHANE         31 https:/… IST     
#>  2 E2023      perGame       008962    HAYES-DAVIS, …        29 https:/… ULK     
#>  3 E2023      perGame       005985    JAMES, MIKE           33 https:/… MCO     
#>  4 E2023      perGame       007831    SHIELDS, SHAV…        29 https:/… MIL     
#>  5 E2023      perGame       009622    BROWN, LORENZO        33 https:/… TEL     
#>  6 E2023      perGame       011204    GRANT, JERIAN         31 https:/… PAN     
#>  7 E2023      perGame       008987    LEDAY, ZACH           29 https:/… PAR     
#>  8 E2023      perGame       009862    PUNTER, KEVIN         30 https:/… PAR     
#>  9 E2023      perGame       012080    LUWAWU-CABARR…        28 https:/… ASV     
#> 10 E2023      perGame       001133    SHENGELIA, TO…        32 https:/… VIR     
#> # ℹ 200 more rows
#> # ℹ 17 more variables: TeamTvCodes <chr>, TeamName <chr>, TeamImageUrl <chr>,
#> #   GP <dbl>, MIN <dbl>, `EFG%` <chr>, `TS%` <chr>, `OREB%` <chr>,
#> #   `DREB%` <chr>, `REB%` <chr>, `AST/TO` <dbl>, `AST-R` <chr>, `TO-R` <chr>,
#> #   `2PA-R` <chr>, `3PA-R` <chr>, `FT-RT` <chr>, Possesions <dbl>