Get player miscellaneous statistics

Usage

getPlayerMisc(season_code)

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

Value

Returns a summary tibble of miscellaneous players statistics for chosen seasons.

Glossary of columns:

Column name	Column extended name	Column description

DD2	Double-doubles	Games with double-digit totals in two of: points, rebounds, assists, steals and blocks

TD3	Triple-doubles	Games with double-digit totals in three of: points, rebounds, assists, steals and blocks

Reference webpage: Stats

Examples

if (FALSE) {

getPlayerMisc(season_code = "E2023")

}