Get player statistics

Usage

getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

statistic_mode

One or more aggregation modes of statistics.
Admitted values are perGame, perMinute and accumulated.

Value

Returns a summary tibble of players statistics for chosen seasons

PTS (Points scored)

2PM (Two-pointers made)

2PA (Two-pointers attempted)

2P% (Two-point %)

3PM (Three-pointers made)

3PA (Three-pointers attempted)

3P% (Three-point %)

FTM (Free throws made)

FTA (Free throws attempted)

FT% (Free-throw %)

OR (Offensive rebounds)

DR (Defensive rebounds)

TR (Total rebounds)

AST (Assists)

STL (Steals)

TO (Turnovers)

BLK (Blocks)

BLKA (Blocks against)

FC (Personal fouls committed)

FD (Personal fouls drawn)

PIR (Performance Index Rating)

Reference webpage: Stats

Examples

getPlayerStats(season_code = "E2023", statistic_mode = "perGame")
#> # A tibble: 214 × 36
#>    SeasonCode StatisticMode Total Ranking Code   Name      Age ImageUrl TeamCode
#>    <chr>      <chr>         <int>   <int> <chr>  <chr>   <int> <chr>    <chr>   
#>  1 E2023      perGame         214       1 007200 LARKIN…    31 https:/… IST     
#>  2 E2023      perGame         214       2 008962 HAYES-…    29 https:/… ULK     
#>  3 E2023      perGame         214       3 005985 JAMES,…    33 https:/… MCO     
#>  4 E2023      perGame         214       4 007831 SHIELD…    29 https:/… MIL     
#>  5 E2023      perGame         214       5 009622 BROWN,…    33 https:/… TEL     
#>  6 E2023      perGame         214       6 011204 GRANT,…    31 https:/… PAN     
#>  7 E2023      perGame         214       7 011225 LEE, P…    28 https:/… ASV     
#>  8 E2023      perGame         214       8 008987 LEDAY,…    29 https:/… PAR     
#>  9 E2023      perGame         214       9 012080 LUWAWU…    28 https:/… ASV     
#> 10 E2023      perGame         214      10 007866 MITOGL…    27 https:/… PAN     
#> # ℹ 204 more rows
#> # ℹ 27 more variables: TeamTvCodes <chr>, TeamName <chr>, TeamImageUrl <chr>,
#> #   GP <dbl>, GS <dbl>, MIN <dbl>, PTS <dbl>, `2FGM` <dbl>, `2FGA` <dbl>,
#> #   `2FG%` <chr>, `3FGM` <dbl>, `3FGA` <dbl>, `3FG%` <chr>, FTM <dbl>,
#> #   FTA <dbl>, `FT%` <chr>, OREB <dbl>, DREB <dbl>, REB <dbl>, AST <dbl>,
#> #   STL <dbl>, TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FR <dbl>, PIR <dbl>