Get team statistics

Usage

getTeamStats(season_code, team_code, phase_type = "All")

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

team_code

One or more team codes as obtained from getCompetitionTeams().
Examples are ASV, MAD, ...

phase_type

One or more phase type codes as obtained from getCompetitionPhases().
Admitted values are RS for regular season, PO for playoffs and FF for final four. Default is All for all.

Value

Returns a list of elements for the chosen teams and seasons:

PlayerAccumulated. Total sum of statistics by player

PlayerAveragePerGame. Average per game of statistics by player

PlayerAveragePer40. Average per 40 minutes of statistics by player

TeamAccumulated. Total sum of statistics of team

TeamAveragePerGame. Average per game of statistics of teams

Glossary of columns:

Column name	Column extended name

GP	Game player

GS	Game started

MIN	Minutes played

PTS	Points scored

2PM	Two-pointers made

2PA	Two-pointers attempted

2P%	Two-point %

3PM	Three-pointers made

3PA	Three-pointers attempted

3FG%	Three-point %

FTM	Free throws made

FTA	Free throws attempted

FT%	Free-throw %

OREB	Offensive rebounds

DREB	Defensive rebounds

TREB	Total rebounds

AST	Assists

STL	Steals

TO	Turnovers

BLK	Blocks

BLKA	Blocks against

FC	Personal fouls committed

FD	Personal fouls drawn

PIR	Performance Index Rating

Reference webpage: Team

Examples

TeamStats = getTeamStats(team_code = "ASV", season_code = c("E2023", "E2022"), phase_type = "RS")

TeamStats$PlayerAccumulated |> head(5)
#> # A tibble: 5 × 36
#>   SeasonCode TeamCode PhaseType Player_ID PlayerName      PlayerAlias TimePlayed
#>   <chr>      <chr>    <chr>     <chr>     <chr>           <chr>            <dbl>
#> 1 E2023      ASV      RS        012729    SCOTT, MIKE     SCOTT, MIKE      35807
#> 2 E2023      ASV      RS        005094    THOMAS, DESHAUN Thomas            6290
#> 3 E2023      ASV      RS        011225    LEE, PARIS      LEE, PARIS       45173
#> 4 E2023      ASV      RS        KIR       KAHUDI, CHARLES Kahudi           23651
#> 5 E2023      ASV      RS        004194    LAUVERGNE, JOF… Lauvergne        37584
#> # ℹ 29 more variables: PIR <dbl>, PTS <dbl>, `2PM` <dbl>, `2PA` <dbl>,
#> #   `3PM` <dbl>, `3PA` <dbl>, FTM <dbl>, FTA <dbl>, FGM <dbl>, FGA <dbl>,
#> #   AM <dbl>, AA <dbl>, REB <dbl>, DREB <dbl>, OREB <dbl>, AST <dbl>,
#> #   STL <dbl>, TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PM <dbl>,
#> #   GP <int>, GS <dbl>, TGS <int>, `3P%` <dbl>, `2P%` <dbl>, `FT%` <dbl>

TeamStats$PlayerAveragePerGame |> head(5)
#> # A tibble: 5 × 36
#>   SeasonCode TeamCode PhaseType Player_ID PlayerName      PlayerAlias TimePlayed
#>   <chr>      <chr>    <chr>     <chr>     <chr>           <chr>            <dbl>
#> 1 E2023      ASV      RS        012729    SCOTT, MIKE     SCOTT, MIKE      1377.
#> 2 E2023      ASV      RS        005094    THOMAS, DESHAUN Thomas           1258 
#> 3 E2023      ASV      RS        011225    LEE, PARIS      LEE, PARIS       1737.
#> 4 E2023      ASV      RS        KIR       KAHUDI, CHARLES Kahudi           1126.
#> 5 E2023      ASV      RS        004194    LAUVERGNE, JOF… Lauvergne        1566 
#> # ℹ 29 more variables: PIR <dbl>, PTS <dbl>, `2PM` <dbl>, `2PA` <dbl>,
#> #   `3PM` <dbl>, `3PA` <dbl>, FTM <dbl>, FTA <dbl>, FGM <dbl>, FGA <dbl>,
#> #   AM <dbl>, AA <dbl>, REB <dbl>, DREB <dbl>, OREB <dbl>, AST <dbl>,
#> #   STL <dbl>, TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PM <dbl>,
#> #   GP <int>, GS <dbl>, TGS <int>, `3P%` <dbl>, `2P%` <dbl>, `FT%` <dbl>

TeamStats$PlayerAveragePer40 |> head(5)
#> # A tibble: 5 × 36
#>   SeasonCode TeamCode PhaseType Player_ID PlayerName      PlayerAlias TimePlayed
#>   <chr>      <chr>    <chr>     <chr>     <chr>           <chr>            <dbl>
#> 1 E2023      ASV      RS        012729    SCOTT, MIKE     SCOTT, MIKE       2400
#> 2 E2023      ASV      RS        005094    THOMAS, DESHAUN Thomas            2400
#> 3 E2023      ASV      RS        011225    LEE, PARIS      LEE, PARIS        2400
#> 4 E2023      ASV      RS        KIR       KAHUDI, CHARLES Kahudi            2400
#> 5 E2023      ASV      RS        004194    LAUVERGNE, JOF… Lauvergne         2400
#> # ℹ 29 more variables: PIR <dbl>, PTS <dbl>, `2PM` <dbl>, `2PA` <dbl>,
#> #   `3PM` <dbl>, `3PA` <dbl>, FTM <dbl>, FTA <dbl>, FGM <dbl>, FGA <dbl>,
#> #   AM <dbl>, AA <dbl>, REB <dbl>, DREB <dbl>, OREB <dbl>, AST <dbl>,
#> #   STL <dbl>, TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PM <dbl>,
#> #   GP <dbl>, GS <dbl>, TGS <dbl>, `3P%` <dbl>, `2P%` <dbl>, `FT%` <dbl>

TeamStats$TeamAccumulated |> head(5)
#> # A tibble: 2 × 33
#>   SeasonCode TeamCode PhaseType TimePlayed   PIR   PTS `2PM` `2PA` `3PM` `3PA`
#>   <chr>      <chr>    <chr>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 E2023      ASV      RS            315000  2209  2037   562  1060   181   563
#> 2 E2022      ASV      RS            408000  2550  2527   671  1253   249   750
#> # ℹ 23 more variables: FTM <dbl>, FTA <dbl>, FGM <dbl>, FGA <dbl>, AM <dbl>,
#> #   AA <dbl>, REB <dbl>, DREB <dbl>, OREB <dbl>, AST <dbl>, STL <dbl>,
#> #   TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PM <dbl>, GP <int>,
#> #   GS <dbl>, TGS <int>, `3P%` <dbl>, `2P%` <dbl>, `FT%` <dbl>

TeamStats$TeamAveragePerGame |> head(5)
#> # A tibble: 2 × 33
#>   SeasonCode TeamCode PhaseType TimePlayed   PIR   PTS `2PM` `2PA` `3PM` `3PA`
#>   <chr>      <chr>    <chr>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 E2023      ASV      RS            12115.    85  78.3  21.6  40.8   7    21.7
#> 2 E2022      ASV      RS            12000     75  74.3  19.7  36.9   7.3  22.1
#> # ℹ 23 more variables: FTM <dbl>, FTA <dbl>, FGM <dbl>, FGA <dbl>, AM <dbl>,
#> #   AA <dbl>, REB <dbl>, DREB <dbl>, OREB <dbl>, AST <dbl>, STL <dbl>,
#> #   TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PM <dbl>, GP <int>,
#> #   GS <dbl>, TGS <int>, `3P%` <dbl>, `2P%` <dbl>, `FT%` <dbl>