Get player statistics

Usage

getPlayerStats(
  season_code,
  statistic_mode = c("perGame", "perMinute", "accumulated")
)

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

statistic_mode

One or more aggregation modes of statistics.
Admitted values are perGame, perMinute and accumulated.

Value

Returns a summary tibble of players statistics for chosen seasons.

Glossary of columns:

Column name	Column extended name

GP	Game player

GS	Game started

MIN	Minutes played

PTS	Points scored

2PM	Two-pointers made

2PA	Two-pointers attempted

2P%	Two-point %

3PM	Three-pointers made

3PA	Three-pointers attempted

3P%	Three-point %

FTM	Free throws made

FTA	Free throws attempted

FT%	Free-throw %

OREB	Offensive rebounds

DREB	Defensive rebounds

TREB	Total rebounds

AST	Assists

STL	Steals

TO	Turnovers

BLK	Blocks

BLKA	Blocks against

FC	Personal fouls committed

FD	Personal fouls drawn

PIR	Performance Index Rating

Reference webpage: Stats

Examples

getPlayerStats(season_code = "E2023", statistic_mode = "perGame")
#> # A tibble: 214 × 34
#>    SeasonCode StatisticMode Player_ID PlayerName     PlayerAge ImageUrl TeamCode
#>    <chr>      <chr>         <chr>     <chr>              <int> <chr>    <chr>   
#>  1 E2023      perGame       007200    LARKIN, SHANE         31 https:/… IST     
#>  2 E2023      perGame       008962    HAYES-DAVIS, …        29 https:/… ULK     
#>  3 E2023      perGame       005985    JAMES, MIKE           33 https:/… MCO     
#>  4 E2023      perGame       007831    SHIELDS, SHAV…        29 https:/… MIL     
#>  5 E2023      perGame       009622    BROWN, LORENZO        33 https:/… TEL     
#>  6 E2023      perGame       011204    GRANT, JERIAN         31 https:/… PAN     
#>  7 E2023      perGame       011225    LEE, PARIS            28 https:/… ASV     
#>  8 E2023      perGame       008987    LEDAY, ZACH           29 https:/… PAR     
#>  9 E2023      perGame       012080    LUWAWU-CABARR…        28 https:/… ASV     
#> 10 E2023      perGame       007866    MITOGLOU, KON…        27 https:/… PAN     
#> # ℹ 204 more rows
#> # ℹ 27 more variables: TeamTvCodes <chr>, TeamName <chr>, TeamImageUrl <chr>,
#> #   GP <dbl>, GS <dbl>, MIN <dbl>, PTS <dbl>, `2PM` <dbl>, `2PA` <dbl>,
#> #   `2P%` <chr>, `3PM` <dbl>, `3PA` <dbl>, `3P%` <chr>, FTM <dbl>, FTA <dbl>,
#> #   `FT%` <chr>, OREB <dbl>, DREB <dbl>, REB <dbl>, AST <dbl>, STL <dbl>,
#> #   TO <dbl>, BLK <dbl>, BLKA <dbl>, FC <dbl>, FD <dbl>, PIR <dbl>