Get team lead statistics

Usage

getTeamLeadStats(season_code, phase_type = "All", subset = "All")

Arguments

season_code

One or more season codes as obtained from getCompetitionHistory().
Examples are E2023 for Euroleague or U2023 for Eurocup 2023.

phase_type

One or more phase type codes as obtained from getCompetitionPhases().
Admitted values are RS for regular season, PO for playoffs and FF for final four. Default is All for all.

subset

One or more game subsets.
Admitted values are HomeGames, AwayGames, GamesWon, GamesLost, ResultsIn5Points (for games resulted in +/-5 points) and All. Default is All.

Value

Returns a list of elements for the chosen seasons, phase_type and subset.

TeamAccumulated. Total sum of statistics of all teams team

TeamAveragePerGame. Average per game of statistics of all teams

Glossary of columns:

Column name	Column extended name

GP	Game player

GS	Game started

MIN	Minutes played

PTS	Points scored

2PM	Two-pointers made

2PA	Two-pointers attempted

2P%	Two-point %

3PM	Three-pointers made

3PA	Three-pointers attempted

3P%	Three-point %

FTM	Free throws made

FTA	Free throws attempted

FT%	Free-throw %

OREB	Offensive rebounds

DREB	Defensive rebounds

TREB	Total rebounds

AST	Assists

STL	Steals

TO	Turnovers

BLK	Blocks

BLKA	Blocks against

FC	Personal fouls committed

FD	Personal fouls drawn

PIR	Performance Index Rating

Reference webpage: TeamLead

Examples

if (FALSE) { # \dontrun{

  TeamLeadStats = getTeamLeadStats(season_code = c("E2022", "E2023"), phase_type = "RS")

  TeamLeadStats$TeamAccumulated |> head(5)

  TeamLeadStats$TeamAveragePerGame |> head(5)

} # }